Adding graceful shutdown via cancellation token
Reorganizing the error structs for ipc specific functionality
