+++ node/src/ipc.rs
@@ -1,14 +1,8 @@
 //! Listens for block notifications and fetches new block templates via IPC
+use crate::error::{classify_error, ErrorKind};
 use tokio::sync::mpsc::Sender;
-use tokio::sync::oneshot;
 pub mod client;
 pub use client::{bytes_to_hex, BitcoinNotification, RequestPriority, SharedBitcoinClient};
-
-enum ErrorKind {
-    Temporary,
-    ConnectionBroken,
-    LogicError,
-}
 
 const MAX_BACKOFF: u64 = 300;
 
@@ -365,67 +359,4 @@ async fn get_template_with_retry(
     } else {
         Err("All attempts failed and no template available".into())
     }
-}
-
-/// Determines if an error indicates a connection/communication failure
-///
-/// This function classifies errors to distinguish between:
-/// * Connection errors: Require reconnection, no point in retrying
-/// * Logic errors: May succeed on retry (temporary issues)
-fn classify_error(error: &Box<dyn std::error::Error>) -> ErrorKind {
-    if let Some(io_err) = error.downcast_ref::<std::io::Error>() {
-        match io_err.kind() {
-            std::io::ErrorKind::ConnectionReset
-            | std::io::ErrorKind::ConnectionAborted
-            | std::io::ErrorKind::BrokenPipe
-            | std::io::ErrorKind::NotConnected => return ErrorKind::ConnectionBroken,
-
-            std::io::ErrorKind::TimedOut
-            | std::io::ErrorKind::Interrupted
-            | std::io::ErrorKind::WouldBlock => return ErrorKind::Temporary,
-
-            _ => {}
-        }
-    }
-
-    if error.downcast_ref::<oneshot::error::RecvError>().is_some() {
-        return ErrorKind::ConnectionBroken;
-    }
-
-    let error_str = error.to_string().to_lowercase();
-
-    if [
-        "connection refused",
-        "connection reset",
-        "connection lost",
-        "broken pipe",
-        "no such file",
-        "permission denied",
-        "disconnected",
-        "bootstrap failed, remote exception",
-        "Method not implemented",
-    ]
-    .iter()
-    .any(|keyword| error_str.contains(keyword))
-    {
-        return ErrorKind::ConnectionBroken;
-    }
-
-    if [
-        "timeout",
-        "try again",
-        "temporary",
-        "interrupted",
-        "busy",
-        "unavailable",
-        "overloaded",
-    ]
-    .iter()
-    .any(|keyword| error_str.contains(keyword))
-    {
-        return ErrorKind::Temporary;
-    }
-
-    // Default to logic error
-    ErrorKind::LogicError
 }
